changed some stuff

+++ routes/update.js
@@ -19,7 +19,7 @@ router.use(express.urlencoded({ extended: true }));
 
 const upload = multer({ storage: fileStorage });
 
-router.put("/item/:param", [authToken, reqLoginTrue, admin ], upload.array("images", 10), async (req, res) => {
+router.put("/item/:param", [authToken, reqLoginTrue, admin ], async (req, res) => {
  let item
  if(req.params.param == "update"){
   item = await Item.updateOne({ _id: req.body.itemId }, 
@@ -34,18 +34,14 @@ router.put("/item/:param", [authToken, reqLoginTrue, admin ], upload.array("imag
     tags: req.body.tags,
     sizes: req.body.sizes,
    }})
-  console.log("item updated", item)
+   return res.send(item)
  }else if(req.params.param == "delete"){
 
   item = await Item.findById(req.body.itemId).select("images")
   if(!item || item.length < 1) return res.status(404)
-  console.log("item", item)
   let itemImgArr = item.images
   
-  console.log("itemImgArr", itemImgArr)
-  
   item = await Item.findByIdAndDelete({_id: req.body.itemId})
-  console.log("Item deleted", item)
 
   let promises = [];
   for(let i = 0; i < itemImgArr.length; i++){
@@ -55,54 +51,72 @@ router.put("/item/:param", [authToken, reqLoginTrue, admin ], upload.array("imag
   Promise.all(promises).then((data)=>{
    // deleteFile(imgPath, req.files)
    console.log("deleted all images from S3 while deleting the item")
-   return res.json(item);
   }).catch((err)=>{
    console.log("error deleting images from S3 while deleting the item", err)
-   return res.send(err.stack);
+  //  return res.send(err.stack);
   }) 
   return res.json(item)
  } else return res.status(500)
 });
 
-router.get("/:id", [authToken, reqLoginTrue, admin ], upload.array("images", 10), async (req, res) => {
+router.get("/:id", [authToken, reqLoginTrue, admin ], async (req, res) => {
  
  let user = await User.findById(req.userId).select("-password")
  if(!user || user.length < 1) return res.status(404).send("Admin not found to Access the page.")
 
  let item = await Item.findById(req.params.id)
  if(!item || item.length < 1) return res.status(404).send("Item not found")
- console.log("item updated", item)
+
  res.render("editItem", {user, item})
 });
 
-router.put("/image/:param", [authToken, reqLoginTrue, admin ], upload.array("images", 10), async (req, res) => {
+router.put("/image/:param?", [authToken, reqLoginTrue, admin ], upload.single("uploadImg"), async (req, res) => {
 
- console.log("req.body", req.body)
- console.log("req.params", req.params)
+ let item
+ if(req.params.param == "delete"){
+   item = await Item.findById(req.body.itemId).select("images")
+   if(!item || item.length < 1) return res.status(400)
+  
+   let imgArr = item.images
+  
+   const index = imgArr.indexOf(req.body.imgId);
+   if (index > -1) {
+    imgArr.splice(index, 1);
+   }
+   
+   await Item.findByIdAndUpdate({_id: req.body.itemId}, 
+   {$set: {
+    "images": imgArr,
+   }})
+   
+   try {
+    await deleteImg(req.body.imgId)
+    console.log("image delete successfully from S3")
+   } catch (error) {
+    console.log("Error Deleting image from S3", error)
+   }
+  res.json(item)
+ }else if(req.params.param == "upload") {
 
- let item = await Item.findById(req.body.itemId).select("images")
- if(!item || item.length < 1) return res.status(400)
-
- let imgArr = item.images
-
- const index = imgArr.indexOf(req.body.imgId);
- console.log("index", index)
- if (index > -1) {
-  imgArr.splice(index, 1);
- }
+  item = await Item.findById(req.query.id).select("images")
+  if(!item || item.length < 1) return res.status(400)
  
- await Item.findByIdAndUpdate({_id: req.body.itemId}, 
- {$set: {
-  "images": imgArr,
- }})
+  item.images.push(req.file.originalname)
  
- try {
-  await deleteImg(req.body.imgId)
-  console.log("image delete successfully from S3")
- } catch (error) {
-  console.log("Error Deleting image from S3", error)
+  item.save().then(()=> {
+   console.log("saved item")
+
+   uploadFile(req.file).then(()=>{
+    console.log("Image Uploaded Successfully");
+    return res.send(item);
+   }).catch((err)=>{
+    console.log(err.stack)
+    return res.send(item);
+   }) 
+
+  })
  }
- res.json(item)
+
 });
 
 module.exports = router;
